+++ lib/src/pluto_dual_grid.dart
@@ -4,13 +4,19 @@ import 'package:pluto_grid/pluto_grid.dart';
 typedef PlutoDualOnSelectedEventCallback = void Function(
     PlutoDualOnSelectedEvent event);
 
+/// In [PlutoDualGrid], set the separation widget between the two grids.
 class PlutoDualGridDivider {
+  /// If [show] is set to true, a separator widget appears between the grids,
+  /// and you can change the width of two grids by dragging them.
   final bool show;
 
+  /// Set the background color.
   final Color backgroundColor;
 
+  /// Set the icon color in the center of the separator widget.
   final Color indicatorColor;
 
+  /// Set the background color when dragging the separator widget.
   final Color draggingColor;
 
   const PlutoDualGridDivider({
@@ -106,7 +112,6 @@ class PlutoDualGridState extends State<PlutoDualGrid> {
           columns: props.columns,
           rows: props.rows,
           columnGroups: props.columnGroups,
-          mode: mode,
           onLoaded: (PlutoGridOnLoadedEvent onLoadedEvent) {
             if (isGridA!) {
               _stateManagerA = onLoadedEvent.stateManager;
@@ -159,9 +164,15 @@ class PlutoDualGridState extends State<PlutoDualGrid> {
           },
           onSorted: props.onSorted,
           onRowChecked: props.onRowChecked,
+          onRowDoubleTap: props.onRowDoubleTap,
+          onRowSecondaryTap: props.onRowSecondaryTap,
+          onRowsMoved: props.onRowsMoved,
           createHeader: props.createHeader,
           createFooter: props.createFooter,
+          rowColorCallback: props.rowColorCallback,
+          columnMenuDelegate: props.columnMenuDelegate,
           configuration: props.configuration,
+          mode: mode,
           key: props.key,
         ),
       ),
@@ -473,25 +484,64 @@ class PlutoDualGridDisplayExpandedAndFixed implements PlutoDualGridDisplay {
 }
 
 class PlutoDualGridProps {
+  /// {@macro pluto_grid_property_columns}
   final List<PlutoColumn> columns;
 
+  /// {@macro pluto_grid_property_rows}
   final List<PlutoRow> rows;
 
+  /// {@macro pluto_grid_property_columnGroups}
   final List<PlutoColumnGroup>? columnGroups;
 
+  /// {@macro pluto_grid_property_onLoaded}
   final PlutoOnLoadedEventCallback? onLoaded;
 
+  /// {@macro pluto_grid_property_onChanged}
   final PlutoOnChangedEventCallback? onChanged;
 
+  /// {@macro pluto_grid_property_onSorted}
   final PlutoOnSortedEventCallback? onSorted;
 
+  /// {@macro pluto_grid_property_onRowChecked}
   final PlutoOnRowCheckedEventCallback? onRowChecked;
 
+  /// {@macro pluto_grid_property_onRowDoubleTap}
+  final PlutoOnRowDoubleTapEventCallback? onRowDoubleTap;
+
+  /// {@macro pluto_grid_property_onRowSecondaryTap}
+  final PlutoOnRowSecondaryTapEventCallback? onRowSecondaryTap;
+
+  /// {@macro pluto_grid_property_onRowsMoved}
+  final PlutoOnRowsMovedEventCallback? onRowsMoved;
+
+  /// {@macro pluto_grid_property_createHeader}
   final CreateHeaderCallBack? createHeader;
 
+  /// {@macro pluto_grid_property_createFooter}
   final CreateFooterCallBack? createFooter;
 
+  /// {@macro pluto_grid_property_rowColorCallback}
+  final PlutoRowColorCallback? rowColorCallback;
+
+  /// {@macro pluto_grid_property_columnMenuDelegate}
+  final PlutoColumnMenuDelegate? columnMenuDelegate;
+
+  /// {@macro pluto_grid_property_configuration}
   final PlutoGridConfiguration configuration;
+
+  /// Execution mode of [PlutoGrid].
+  ///
+  /// [PlutoGridMode.normal]
+  /// {@macro pluto_grid_mode_normal}
+  ///
+  /// [PlutoGridMode.select], [PlutoGridMode.selectWithOneTap]
+  /// {@macro pluto_grid_mode_select}
+  ///
+  /// [PlutoGridMode.popup]
+  /// {@macro pluto_grid_mode_popup}
+  final PlutoGridMode? mode;
+
+  final Key? key;
 
   const PlutoDualGridProps({
     required this.columns,
@@ -501,38 +551,65 @@ class PlutoDualGridProps {
     this.onChanged,
     this.onSorted,
     this.onRowChecked,
+    this.onRowDoubleTap,
+    this.onRowSecondaryTap,
+    this.onRowsMoved,
     this.createHeader,
     this.createFooter,
+    this.rowColorCallback,
+    this.columnMenuDelegate,
     this.configuration = const PlutoGridConfiguration(),
+    this.mode,
     this.key,
   });
-
-  final Key? key;
 
   PlutoDualGridProps copyWith({
     List<PlutoColumn>? columns,
     List<PlutoRow>? rows,
-    List<PlutoColumnGroup>? columnGroups,
-    PlutoOnLoadedEventCallback? onLoaded,
-    PlutoOnChangedEventCallback? onChanged,
-    PlutoOnSortedEventCallback? onSorted,
-    PlutoOnRowCheckedEventCallback? onRowChecked,
-    CreateHeaderCallBack? createHeader,
-    CreateFooterCallBack? createFooter,
+    PlutoOptional<List<PlutoColumnGroup>?>? columnGroups,
+    PlutoOptional<PlutoOnLoadedEventCallback?>? onLoaded,
+    PlutoOptional<PlutoOnChangedEventCallback?>? onChanged,
+    PlutoOptional<PlutoOnSortedEventCallback?>? onSorted,
+    PlutoOptional<PlutoOnRowCheckedEventCallback?>? onRowChecked,
+    PlutoOptional<PlutoOnRowDoubleTapEventCallback?>? onRowDoubleTap,
+    PlutoOptional<PlutoOnRowSecondaryTapEventCallback?>? onRowSecondaryTap,
+    PlutoOptional<PlutoOnRowsMovedEventCallback?>? onRowsMoved,
+    PlutoOptional<CreateHeaderCallBack?>? createHeader,
+    PlutoOptional<CreateFooterCallBack?>? createFooter,
+    PlutoOptional<PlutoRowColorCallback?>? rowColorCallback,
+    PlutoOptional<PlutoColumnMenuDelegate?>? columnMenuDelegate,
     PlutoGridConfiguration? configuration,
+    PlutoOptional<PlutoGridMode?>? mode,
     Key? key,
   }) {
     return PlutoDualGridProps(
       columns: columns ?? this.columns,
       rows: rows ?? this.rows,
-      columnGroups: columnGroups ?? this.columnGroups,
-      onLoaded: onLoaded ?? this.onLoaded,
-      onChanged: onChanged ?? this.onChanged,
-      onSorted: onSorted ?? this.onSorted,
-      onRowChecked: onRowChecked ?? this.onRowChecked,
-      createHeader: createHeader ?? this.createHeader,
-      createFooter: createFooter ?? this.createFooter,
+      columnGroups:
+          columnGroups == null ? this.columnGroups : columnGroups.value,
+      onLoaded: onLoaded == null ? this.onLoaded : onLoaded.value,
+      onChanged: onChanged == null ? this.onChanged : onChanged.value,
+      onSorted: onSorted == null ? this.onSorted : onSorted.value,
+      onRowChecked:
+          onRowChecked == null ? this.onRowChecked : onRowChecked.value,
+      onRowDoubleTap:
+          onRowDoubleTap == null ? this.onRowDoubleTap : onRowDoubleTap.value,
+      onRowSecondaryTap: onRowSecondaryTap == null
+          ? this.onRowSecondaryTap
+          : onRowSecondaryTap.value,
+      onRowsMoved: onRowsMoved == null ? this.onRowsMoved : onRowsMoved.value,
+      createHeader:
+          createHeader == null ? this.createHeader : createHeader.value,
+      createFooter:
+          createFooter == null ? this.createFooter : createFooter.value,
+      rowColorCallback: rowColorCallback == null
+          ? this.rowColorCallback
+          : rowColorCallback.value,
+      columnMenuDelegate: columnMenuDelegate == null
+          ? this.columnMenuDelegate
+          : columnMenuDelegate.value,
       configuration: configuration ?? this.configuration,
+      mode: mode == null ? this.mode : mode.value,
       key: key ?? this.key,
     );
   }